// cli/src/config.d.ts
export declare const config: {
    database: {
        url: string;
    };
    redis: {
        url: string;
    };
    hedera: {
        mirrorNodeUrl: {
            mainnet: string;
            testnet: string;
        };
    };
    cache: {
        refreshInterval: number;
    };
};
//# sourceMappingURL=config.d.ts.map